# Updated unit animation code. Added UnitAnimation class, to act as the interface between the entity and actor. (Currently doesn't work very well, but it does make animations loop smoothly and sometimes kind of makes them stay synchronised.) Fixed corpse animation - it now plays the final frame of the death animation before turning static. Fixed update/interpolate timings. Added JS function saveProfileData. Updated ffmpeg library.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4880 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/ObjectBase.h
@@ -26,8 +26,8 @@ public:
 		float m_Speed;
 		// fraction of the way through the animation that the interesting bit(s)
 		// happens 
-		double m_ActionPos;
-		double m_ActionPos2;
+		float m_ActionPos;
+		float m_ActionPos2;
 	};
 
 	struct Prop {
